+++ internal/loadbalancer/server.go
@@ -5,7 +5,9 @@ import (
 	"github.com/NodeFactoryIo/vedran/internal/auth"
 	"github.com/NodeFactoryIo/vedran/internal/configuration"
 	"github.com/NodeFactoryIo/vedran/internal/models"
+	"github.com/NodeFactoryIo/vedran/internal/repositories"
 	"github.com/NodeFactoryIo/vedran/internal/router"
+	"github.com/NodeFactoryIo/vedran/internal/schedule/checkactive"
 	"github.com/asdine/storm/v3"
 	log "github.com/sirupsen/logrus"
 	"net/http"
@@ -29,6 +31,21 @@ func StartLoadBalancerServer(props configuration.Configuration) {
 	}
 	log.Debug("Successfully connected to database")
 
+	// initialize repos
+	var repos = &repositories.Repos{}
+	repos.PingRepo = repositories.NewPingRepo(database)
+	repos.MetricsRepo = repositories.NewMetricsRepo(database)
+	repos.RecordRepo = repositories.NewRecordRepo(database)
+	repos.NodeRepo = repositories.NewNodeRepo(database)
+	err = repos.NodeRepo.InitNodeRepo()
+	if err != nil {
+		log.Fatalf("Failed initializing node repo because of: %v", err)
+	}
+
+	// starts task that checks active nodes
+	checkactive.StartScheduledTask(repos)
+
+	// whitelist
 	whitelistEnabled := len(props.Whitelist) > 0
 	// save whitelisted id-s
 	if whitelistEnabled {
@@ -46,7 +63,7 @@ func StartLoadBalancerServer(props configuration.Configuration) {
 
 	// start server
 	log.Infof("Starting vedran load balancer on port :%d...", props.Port)
-	r := router.CreateNewApiRouter(database, whitelistEnabled)
+	r := router.CreateNewApiRouter(*repos, whitelistEnabled)
 	err = http.ListenAndServe(fmt.Sprintf(":%d", props.Port), r)
 	if err != nil {
 		log.Error(err)